Add testutil package
